[RSTC] Regenerated demo for rstc_rstc_reset_cause_sam_e54_xpro
MH3-79204

+++ apps/rstc/rstc_reset_cause/firmware/src/config/sam_e54_xpro/peripheral/wdt/plib_wdt.c
@@ -93,19 +93,19 @@ void WDT_Disable( void )
 
 void WDT_EnableWindowMode( void )
 {
-	while(WDT_REGS->WDT_SYNCBUSY != 0U)
+    while(WDT_REGS->WDT_SYNCBUSY != 0U)
     {
 
     }
-	
+
     /* Window mode can be changed only if peripheral is disabled or ALWAYS ON bit is set */
     if(((WDT_REGS->WDT_CTRLA & WDT_CTRLA_ENABLE_Msk) == 0U) || ((WDT_REGS->WDT_CTRLA & WDT_CTRLA_ALWAYSON_Msk) != 0U))
     {
         /* Enable window mode */
         WDT_REGS->WDT_CTRLA |= (uint8_t)WDT_CTRLA_WEN_Msk;
     }
-	
-	while(WDT_REGS->WDT_SYNCBUSY != 0U)
+
+    while(WDT_REGS->WDT_SYNCBUSY != 0U)
     {
 
     }
@@ -113,19 +113,19 @@ void WDT_EnableWindowMode( void )
 
 void WDT_DisableWindowMode( void )
 {
-	while(WDT_REGS->WDT_SYNCBUSY != 0U)
+    while(WDT_REGS->WDT_SYNCBUSY != 0U)
     {
 
     }
-	
+
     /* Window mode can be changed only if peripheral is disabled or ALWAYS ON bit is set */
     if(((WDT_REGS->WDT_CTRLA & WDT_CTRLA_ENABLE_Msk) == 0U) || ((WDT_REGS->WDT_CTRLA & WDT_CTRLA_ALWAYSON_Msk) != 0U))
     {
         /* Disable window mode */
         WDT_REGS->WDT_CTRLA &= (uint8_t)(~WDT_CTRLA_WEN_Msk);
     }
-	
-	while(WDT_REGS->WDT_SYNCBUSY != 0U)
+
+    while(WDT_REGS->WDT_SYNCBUSY != 0U)
     {
 
     }
@@ -150,7 +150,7 @@ bool WDT_IsWindowModeEnabled(void)
 void WDT_TimeoutPeriodSet(uint8_t TimeoutPeriod)
 {
     /* Set WDT timeout period */
-    WDT_REGS->WDT_CONFIG = (WDT_REGS->WDT_CONFIG & ~WDT_CONFIG_PER_Msk) | (TimeoutPeriod & WDT_CONFIG_PER_Msk);
+    WDT_REGS->WDT_CONFIG = (WDT_REGS->WDT_CONFIG & (uint8_t)~WDT_CONFIG_PER_Msk) | (TimeoutPeriod & (uint8_t)WDT_CONFIG_PER_Msk);
 }
 
 /* If application intends to stay in active mode after clearing WDT, then use WDT_Clear API to clear the WDT. This avoids CPU from waiting or stalling for Synchronization.
